# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmritiSetu - Ancestor Remembrance App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # USERS
...
  /ancestors:
    get:
      operationId: listAncestors
      tags: [ancestors]
      summary: List all ancestors for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: familySide
          in: query
          schema:
            type: ["string", "null"]
            enum: [Maternal, Paternal, null]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of ancestors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ancestor"

    post:
      operationId: createAncestor
      tags: [ancestors]
      summary: Add a new ancestor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAncestorBody"
      responses:
        "201":
          description: Ancestor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ancestor"
  /ancestors/{id}:
    get:
      operationId: getAncestor
      tags: [ancestors]
      summary: Get a specific ancestor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ancestor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ancestor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateAncestor
      tags: [ancestors]
      summary: Update ancestor details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAncestorBody"
      responses:
        "200":
          description: Updated ancestor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ancestor"

    delete:
      operationId: deleteAncestor
      tags: [ancestors]
      summary: Delete an ancestor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # REMINDERS
...
  /ai/remembrance-message:
    post:
      operationId: generateRemembranceMessage
      tags: [ai]
      summary: Generate a personalized remembrance message for an ancestor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMessageBody"
      responses:
        "200":
          description: Generated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedMessage"
  /ai/ritual-suggestions:
    post:
      operationId: getRitualSuggestions
      tags: [ai]
      summary: Get AI-suggested rituals for an ancestor's remembrance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RitualSuggestionsBody"
      responses:
        "200":
          description: Ritual suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RitualSuggestions"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        language:
          type: string
          enum: [en, te, hi]
          default: en
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        language:
          type: string
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - language
        - createdAt
...
    Ancestor:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        fullName:
          type: string
        relationship:
          type: string
        gender:
          type: string
          enum: [Male, Female, Other]
        photoUrl:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        dateOfDeath:
          type: string
        placeOfDeath:
          type: ["string", "null"]
        tithi:
          type: ["string", "null"]
        nakshatram:
          type: ["string", "null"]
        masam:
          type: ["string", "null"]
        paksham:
          type: ["string", "null"]
          enum: [Shukla, Krishna, null]
        samvatsaram:
          type: ["string", "null"]
        teluguYearName:
          type: ["string", "null"]
        timeOfDeath:
          type: ["string", "null"]
        familySide:
          type: ["string", "null"]
          enum: [Maternal, Paternal, null]
        gotram:
          type: ["string", "null"]
        reminderDaysBefore:
          type: integer
          default: 7
        ritualNotes:
          type: ["string", "null"]
        traditions:
          type: ["string", "null"]
        favoriteMemories:
          type: ["string", "null"]
        prasadamDetails:
          type: ["string", "null"]
        daysUntilNext:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - fullName
        - relationship
        - gender
        - dateOfDeath
        - reminderDaysBefore
        - createdAt
    CreateAncestorBody:
      type: object
      properties:
        fullName:
          type: string
        relationship:
          type: string
        gender:
          type: string
          enum: [Male, Female, Other]
        photoUrl:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        dateOfDeath:
          type: string
        placeOfDeath:
          type: ["string", "null"]
        tithi:
          type: ["string", "null"]
        nakshatram:
          type: ["string", "null"]
        masam:
          type: ["string", "null"]
        paksham:
          type: ["string", "null"]
        samvatsaram:
          type: ["string", "null"]
        teluguYearName:
          type: ["string", "null"]
        timeOfDeath:
          type: ["string", "null"]
        familySide:
          type: ["string", "null"]
        gotram:
          type: ["string", "null"]
        reminderDaysBefore:
          type: integer
          default: 7
        ritualNotes:
          type: ["string", "null"]
        traditions:
          type: ["string", "null"]
        favoriteMemories:
          type: ["string", "null"]
        prasadamDetails:
          type: ["string", "null"]
      required:
        - fullName
        - relationship
        - gender
        - dateOfDeath
    UpdateAncestorBody:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        relationship:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        dateOfDeath:
          type: ["string", "null"]
        placeOfDeath:
          type: ["string", "null"]
        tithi:
          type: ["string", "null"]
        nakshatram:
          type: ["string", "null"]
        masam:
          type: ["string", "null"]
        paksham:
          type: ["string", "null"]
        samvatsaram:
          type: ["string", "null"]
        teluguYearName:
          type: ["string", "null"]
        timeOfDeath:
          type: ["string", "null"]
        familySide:
          type: ["string", "null"]
        gotram:
          type: ["string", "null"]
        reminderDaysBefore:
          type: ["integer", "null"]
        ritualNotes:
          type: ["string", "null"]
        traditions:
          type: ["string", "null"]
        favoriteMemories:
          type: ["string", "null"]
        prasadamDetails:
          type: ["string", "null"]

    # REMINDERS
...
    GenerateMessageBody:
      type: object
      properties:
        ancestorId:
          type: integer
        occasion:
          type: string
          enum: [anniversary, birthday, general]
      required:
        - ancestorId
        - occasion
    GeneratedMessage:
      type: object
      properties:
        message:
          type: string
        language:
          type: string
      required:
        - message
        - language
    RitualSuggestionsBody:
      type: object
      properties:
        ancestorId:
          type: integer
      required:
        - ancestorId
    RitualSuggestions:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: string
        foodOfferings:
          type: array
          items:
            type: string
        prayers:
          type: array
          items:
            type: string
      required:
        - suggestions
        - foodOfferings
        - prayers