# lib/api-zod/src/generated/types/createAncestorBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmritiSetu - Ancestor Remembrance App API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAncestorBodyGender } from "./createAncestorBodyGender";

export interface CreateAncestorBody {
  fullName: string;
  relationship: string;
  gender: CreateAncestorBodyGender;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  dateOfDeath: string;
  /** @nullable */
  placeOfDeath?: string | null;
  /** @nullable */
  tithi?: string | null;
  /** @nullable */
  nakshatram?: string | null;
  /** @nullable */
  masam?: string | null;
  /** @nullable */
  paksham?: string | null;
  /** @nullable */
  samvatsaram?: string | null;
  /** @nullable */
  teluguYearName?: string | null;
  /** @nullable */
  timeOfDeath?: string | null;
  /** @nullable */
  familySide?: string | null;
  /** @nullable */
  gotram?: string | null;
  reminderDaysBefore?: number;
  /** @nullable */
  ritualNotes?: string | null;
  /** @nullable */
  traditions?: string | null;
  /** @nullable */
  favoriteMemories?: string | null;
  /** @nullable */
  prasadamDetails?: string | null;
}