# lib/api-zod/src/generated/types/updateAncestorBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmritiSetu - Ancestor Remembrance App API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAncestorBody {
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  relationship?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  dateOfDeath?: string | null;
  /** @nullable */
  placeOfDeath?: string | null;
  /** @nullable */
  tithi?: string | null;
  /** @nullable */
  nakshatram?: string | null;
  /** @nullable */
  masam?: string | null;
  /** @nullable */
  paksham?: string | null;
  /** @nullable */
  samvatsaram?: string | null;
  /** @nullable */
  teluguYearName?: string | null;
  /** @nullable */
  timeOfDeath?: string | null;
  /** @nullable */
  familySide?: string | null;
  /** @nullable */
  gotram?: string | null;
  /** @nullable */
  reminderDaysBefore?: number | null;
  /** @nullable */
  ritualNotes?: string | null;
  /** @nullable */
  traditions?: string | null;
  /** @nullable */
  favoriteMemories?: string | null;
  /** @nullable */
  prasadamDetails?: string | null;
}