# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmritiSetu - Ancestor Remembrance App API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  language: string;
  /** @nullable */
  photoUrl?: string | null;
  createdAt: string;
}